@@ -1,3 +1,4 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using TILApp.Models;
@@ -17,29 +18,27 @@ namespace TILApp.Controllers
 
         // GET: api/User
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<UserDto>>> GetUser()
+        public async Task<ActionResult<IEnumerable<User.Public>>> GetUser()
         {
-            if (db.User == null) return NotFound(); 
+            if (db.User == null) return NotFound();
 
-            return UserDto.convertedFrom(await db.User.ToListAsync());
+            return new User.Public().List(await db.User.ToListAsync());
         }
 
         // GET: api/User/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<UserDto>> GetUser(int id)
+        public async Task<ActionResult<User.Public>> GetUser(int id)
         {
             if (db.User == null) return NotFound(); 
 
             var user = await db.User.FindAsync(id);
 
-            if (user == null) return NotFound(); 
-
-            return UserDto.convertedFrom(user);
+            return (user != null) ? new User.Public(user) : NotFound();
         }
 
         // GET: api/User/5/Acronyms
         [HttpGet("{id}/Acronyms")]
-        public async Task<ActionResult<IEnumerable<AcronymDto>>> GetAcronyms(int id)
+        public async Task<ActionResult<IEnumerable<Acronym.Dto>>> GetAcronyms(string? id)
         {
             if (db.User == null) return NotFound(); 
 
@@ -47,13 +46,13 @@ namespace TILApp.Controllers
 
             if (user == null || user.Acronyms == null) return NotFound();
 
-            return AcronymDto.convertedFrom(user.Acronyms.ToList());
+            return new Acronym.Dto().List(user.Acronyms.ToList());
         }
 
         // PUT: api/User/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPut("{id}")]
-        public async Task<IActionResult> PutUser(int id, UserDto dto)
+        [HttpPut("{id}"), Authorize]
+        public async Task<IActionResult> PutUser(string? id, User.Public dto)
         {
             var user = await db.User.Where(i => i.Id == id).FirstAsync();
 
@@ -71,27 +70,29 @@ namespace TILApp.Controllers
                 else throw;
             }
 
-            return Ok(UserDto.convertedFrom(user));
+            return Ok(new User.Public(user));
         }
 
         // POST: api/User
+        // POST is disabled; new Users can only be added through the identity api end point /register
+        
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPost]
-        public async Task<ActionResult<UserDto>> PostUser(UserDto dto)
-        {
-            if (db.User == null) return Problem("Entity set 'AcronymContext.User'  is null.");
-
-            var user = new User() { Name = dto.Name, UserName = dto.UserName };
-
-            db.User.Add(user);
-            await db.SaveChangesAsync();
-
-            return CreatedAtAction("GetUser", new { id = dto.Id }, dto);
-        }
+        // [HttpPost]
+        // public async Task<ActionResult<User.Public>> PostUser(User.Public publicUser)
+        // {
+        //     if (db.User == null) return Problem("Entity set 'AcronymContext.User'  is null.");
+        //
+        //     var user = new User() { Name = publicUser.Name, UserName = publicUser.UserName };
+        //
+        //     db.User.Add(user);
+        //     await db.SaveChangesAsync();
+        //
+        //     return CreatedAtAction("GetUser", new { id = publicUser.Id }, publicUser);
+        // }
 
         // DELETE: api/User/5
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteUser(int id)
+        [HttpDelete("{id}"), Authorize]
+        public async Task<IActionResult> DeleteUser(string id)
         {
             if (db.User == null) return NotFound(); 
 
@@ -105,7 +106,7 @@ namespace TILApp.Controllers
             return NoContent();
         }
 
-        private bool UserExists(int id)
+        private bool UserExists(string? id)
         {
             return (db.User?.Any(e => e.Id == id)).GetValueOrDefault();
         }
