@@ -1,28 +1,28 @@
-﻿namespace TILApp.Models
+﻿namespace TILApp.Models;
+
+public class Category
 {
+    public int Id { get; set; }
+    public string? Name { get; set; }
 
-    public class Category
+    public ICollection<Acronym>? Acronyms { get; set; }
+
+    public class CDto
     {
         public int Id { get; set; }
         public string? Name { get; set; }
 
-        public ICollection<Acronym>? Acronyms { get; set; }
-    }
+        public CDto() { }
 
-    public class CategoryDto
-    {
-        public int Id { get; set; }
-        public string? Name { get; set; }
-
-        public static CategoryDto convertedFrom(Category category)
+        public CDto(Category category)
         {
-            return new CategoryDto() { Id = category.Id, Name = category.Name };
+            Id = category.Id;
+            Name = category.Name;
         }
 
-        public static List<CategoryDto> convertedFrom(List<Category> categories)
+        public List<CDto> List(List<Category> categories)
         {
-            return categories.Select(a => convertedFrom(a)).OrderBy(i => i.Id).ToList();
+            return categories.Select(a => new CDto(a)).OrderBy(i => i.Id).ToList();
         }
     }
-
 }
