@@ -7,28 +7,33 @@ namespace TILApp.Models
         public string? Short { get; set; }
         public string? Long { get; set; }
 
-        public User? User { get; set; }
-        public int? UserId { get; set; }
+        public User? User { get; set; } 
+        public string? UserId { get; set; }
 
         public ICollection<Category>? Categories { get; set; }
-    }
 
-    public class AcronymDto
-    {
-        public int Id { get; set; }
-        public string? Short { get; set; }
-        public string? Long { get; set; }
-        public int? UserId { get; set; }
-
-        public static AcronymDto convertedFrom(Acronym acronym)
+        public class Dto
         {
-            return new AcronymDto() { Id = acronym.Id, Long = acronym.Long, Short = acronym.Short, UserId = acronym.UserId };
-        }
+            public int Id { get; set; }
+            public string? Short { get; set; }
+            public string? Long { get; set; }
+            public string? UserId { get; set; }
 
-        public static List<AcronymDto> convertedFrom(List<Acronym> acronyms)
-        {
-            return acronyms.Select(a => convertedFrom(a)).OrderBy(i => i.Id).ToList();
+            public Dto() { }
+
+            public Dto(Acronym acronym)
+            {
+                Id = acronym.Id;
+                Short = acronym.Short;
+                Long = acronym.Long;
+                UserId = acronym.UserId;
+            }
+
+            public List<Dto> List(List<Acronym> acronyms)
+            {
+                return acronyms.Select(a => new Dto(a)).OrderBy(i => i.Id).ToList();
+            }
         }
     }
-
 }
+
